refactor: Functional Interface 제거 #204
