+++ backend/src/main/java/ddangkong/facade/room/balance/roomvote/RoomBalanceVoteFacade.java
@@ -8,13 +8,12 @@ import ddangkong.domain.room.member.Member;
 import ddangkong.domain.room.member.RoomMembers;
 import ddangkong.exception.BadRequestException;
 import ddangkong.facade.balance.vote.dto.ContentTotalBalanceVoteResponse;
+import ddangkong.facade.room.balance.roomvote.context.VoteContext;
 import ddangkong.facade.room.balance.roomvote.dto.ContentRoomBalanceVoteResponse;
 import ddangkong.facade.room.balance.roomvote.dto.RoomBalanceVoteRequest;
 import ddangkong.facade.room.balance.roomvote.dto.RoomBalanceVoteResponse;
 import ddangkong.facade.room.balance.roomvote.dto.RoomBalanceVoteResultResponse;
 import ddangkong.facade.room.balance.roomvote.dto.VoteFinishedResponse;
-import ddangkong.facade.room.balance.roomvote.operation.VoteContext;
-import ddangkong.facade.room.balance.roomvote.operation.VoteOperation;
 import ddangkong.service.balance.content.BalanceContentService;
 import ddangkong.service.balance.option.BalanceOptionService;
 import ddangkong.service.balance.vote.TotalBalanceVoteService;
@@ -47,39 +46,26 @@ public class RoomBalanceVoteFacade {
 
     @Transactional
     public RoomBalanceVoteResponse createVote(RoomBalanceVoteRequest request, Long roomId, Long contentId) {
-        return handleVoteOperation(roomId, contentId, voteContext -> {
-            if (voteContext.isVoteFinished()) {
-                throw new BadRequestException("이미 투표가 종료되었습니다.");
-            }
-            Member member = voteContext.getRoomMembers().getMember(request.memberId());
-            RoomBalanceVote roomBalanceVote = roomBalanceVoteService.createVote(member, voteContext.getBalanceOptions(),
-                    request.optionId());
-            return new RoomBalanceVoteResponse(roomBalanceVote);
-        });
+        VoteContext voteContext = getVoteContext(roomId, contentId);
+        if (voteContext.isVoteFinished()) {
+            throw new BadRequestException("이미 투표가 종료되었습니다.");
+        }
+        Member member = voteContext.getRoomMembers().getMember(request.memberId());
+        RoomBalanceVote roomBalanceVote = roomBalanceVoteService.createVote(member, voteContext.getBalanceOptions(),
+                request.optionId());
+        return new RoomBalanceVoteResponse(roomBalanceVote);
     }
 
     @Transactional(readOnly = true)
     public RoomBalanceVoteResultResponse getAllVoteResult(Long roomId, Long contentId) {
-        return handleVoteOperation(roomId, contentId, voteContext -> {
-            if (voteContext.isVoteNotFinished()) {
-                throw new BadRequestException("투표가 끝나지 않아 투표 결과를 조회할 수 없습니다.");
-            }
-            ContentRoomBalanceVoteResponse group = getContentRoomBalanceVoteResponse(voteContext.getRoomMembers(),
-                    voteContext.getBalanceOptions());
-            ContentTotalBalanceVoteResponse total = getContentTotalBalanceVoteResponse(voteContext.getBalanceOptions());
-            return new RoomBalanceVoteResultResponse(group, total);
-        });
-    }
-
-    private <T> T handleVoteOperation(Long roomId, Long contentId, VoteOperation<T> voteOperation) {
-        Room room = roomService.getRoom(roomId);
-        BalanceContent balanceContent = balanceContentService.getBalanceContent(contentId);
-        RoomMembers roomMembers = memberService.findRoomMembers(room);
-        BalanceOptions balanceOptions = balanceOptionService.getBalanceOptions(balanceContent);
-        boolean roundFinished = roomContentService.isRoundFinished(room, balanceContent);
-        boolean voteFinished = roomBalanceVoteService.isVoteFinished(roomMembers, balanceOptions);
-
-        return voteOperation.execute(new VoteContext(roomMembers, balanceOptions, roundFinished || voteFinished));
+        VoteContext voteContext = getVoteContext(roomId, contentId);
+        if (voteContext.isVoteNotFinished()) {
+            throw new BadRequestException("투표가 끝나지 않아 투표 결과를 조회할 수 없습니다.");
+        }
+        ContentRoomBalanceVoteResponse group = getContentRoomBalanceVoteResponse(voteContext.getRoomMembers(),
+                voteContext.getBalanceOptions());
+        ContentTotalBalanceVoteResponse total = getContentTotalBalanceVoteResponse(voteContext.getBalanceOptions());
+        return new RoomBalanceVoteResultResponse(group, total);
     }
 
     private ContentRoomBalanceVoteResponse getContentRoomBalanceVoteResponse(RoomMembers roomMembers,
@@ -100,9 +86,19 @@ public class RoomBalanceVoteFacade {
 
     @Transactional(readOnly = true)
     public VoteFinishedResponse getVoteFinished(Long roomId, Long contentId) {
-        return handleVoteOperation(roomId, contentId, voteContext -> {
-            Member master = voteContext.getRoomMembers().getMaster();
-            return new VoteFinishedResponse(voteContext.isVoteFinished(), master);
-        });
+        VoteContext voteContext = getVoteContext(roomId, contentId);
+        Member master = voteContext.getRoomMembers().getMaster();
+        return new VoteFinishedResponse(voteContext.isVoteFinished(), master);
+    }
+
+    private VoteContext getVoteContext(Long roomId, Long contentId) {
+        Room room = roomService.getRoom(roomId);
+        BalanceContent balanceContent = balanceContentService.getBalanceContent(contentId);
+        RoomMembers roomMembers = memberService.findRoomMembers(room);
+        BalanceOptions balanceOptions = balanceOptionService.getBalanceOptions(balanceContent);
+        boolean roundFinished = roomContentService.isRoundFinished(room, balanceContent);
+        boolean voteFinished = roomBalanceVoteService.isVoteFinished(roomMembers, balanceOptions);
+
+        return new VoteContext(roomMembers, balanceOptions, roundFinished || voteFinished);
     }
 }
