--- backend/src/main/java/ddangkong/facade/room/balance/roomvote/operation/VoteOperation.java
@@ -1,7 +1,0 @@
-package ddangkong.facade.room.balance.roomvote.operation;
-
-@FunctionalInterface
-public interface VoteOperation<T> {
-
-    T execute(VoteContext voteContext);
-}
